# client/json/swagger2.json
{"swagger":"2.0","info":{"description":"","version":"1.1.0","title":"userswebapi","contact":{"name":"Ricardo Faria","email":"ricardo.faria@outlook.com.br"}},"host":"userswebapi.apispark.net","basePath":"/v1","schemes":["https"],"paths":{"/users/":{"get":{"tags":["Usuarios"],"summary":"Loads a list of User","description":"Loads a list of User","operationId":"getUserList","consumes":[],"produces":["application/json","application/x-yaml","application/xml","text/xml"],"parameters":[{"name":"email","in":"query","description":"Allows to filter the collections of result by the value of field email","required":false,"type":"string"},{"name":"id","in":"query","description":"Allows to filter the collections of result by the value of field id","required":false,"type":"string"},{"name":"password","in":"query","description":"Allows to filter the collections of result by the value of field password","required":false,"type":"string"},{"name":"$size","in":"query","description":"Size of the page to retrieve. Integer value","required":false,"type":"string"},{"name":"login","in":"query","description":"Allows to filter the collections of result by the value of field login","required":false,"type":"string"},{"name":"$sort","in":"query","description":"Order in which to retrieve the results. Mutliple sort criteria can be passed. Example: sort=age ASC,height DESC","required":false,"type":"string"},{"name":"$page","in":"query","description":"Number of the page to retrieve. Integer value.","required":false,"type":"string"},{"name":"name","in":"query","description":"Allows to filter the collections of result by the value of field name","required":false,"type":"string"}],"responses":{"200":{"description":"200 status response","schema":{"type":"array","items":{"$ref":"#/definitions/User"}}},"400":{"description":"400 status response"}}},"post":{"tags":["Usuarios"],"summary":"Adds a User","description":"Adds a User","operationId":"postUserList","consumes":["application/json","application/x-yaml","application/xml","text/xml"],"produces":["application/json","application/x-yaml","application/xml","text/xml"],"parameters":[{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"200 status response","schema":{"$ref":"#/definitions/User"}}}}},"/users/{userid}":{"get":{"tags":["Usuarios"],"summary":"Loads a User","description":"Loads a User","operationId":"getUser","consumes":[],"produces":["application/json","application/x-yaml","application/xml","text/xml"],"parameters":[{"name":"userid","in":"path","description":"Identifier of the User","required":true,"type":"string"}],"responses":{"200":{"description":"200 status response","schema":{"$ref":"#/definitions/User"}},"400":{"description":"400 status response"}}},"put":{"tags":["Usuarios"],"summary":"Stores a User","description":"Stores a User","operationId":"putUser","consumes":["application/json","application/x-yaml","application/xml","text/xml"],"produces":["application/json","application/x-yaml","application/xml","text/xml"],"parameters":[{"name":"userid","in":"path","description":"Identifier of the User","required":true,"type":"string"},{"in":"body","name":"body","required":false,"schema":{"$ref":"#/definitions/User"}}],"responses":{"200":{"description":"200 status response","schema":{"$ref":"#/definitions/User"}}}},"delete":{"tags":["Usuarios"],"summary":"Deletes a User","description":"Deletes a User","operationId":"deleteUser","consumes":[],"produces":[],"parameters":[{"name":"userid","in":"path","description":"Identifier of the User","required":true,"type":"string"}],"responses":{"200":{"description":"200 status response"}}}}},"securityDefinitions":{"HTTP_BASIC":{"type":"basic"}},"definitions":{"User":{"properties":{"id":{"type":"string","example":"sample id","description":"Auto-generated primary key field"},"name":{"type":"string","example":"sample name"},"login":{"type":"string","example":"sample login"},"email":{"type":"string","example":"sample email"},"password":{"type":"string","example":"sample password"}}}}}